% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_webpage.R
\name{add_dependency}
\alias{add_dependency}
\title{Add HTML dependency}
\usage{
add_dependency(root_dir, ...)
}
\arguments{
\item{root_dir}{Path to the webpage root directory}

\item{...}{Path to the dependency}
}
\description{
Copies an HTML dependency from the package template into the current working
directory.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
add_dependency("css", "styles.css")
\dontshow{\}) # examplesIf}
}
