% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_webpage.R
\name{build_read_path_rel}
\alias{build_read_path_rel}
\title{Build relative path to dependency}
\usage{
build_read_path_rel(root_dir, write_path)
}
\arguments{
\item{root_dir}{The path to the root directory of the webpage.}

\item{write_path}{The write path of the dependency.}
}
\value{
The relative path of the dependency.
}
\description{
Builds the path to the dependency relative to the installed package files.
}
\details{
Starts with the write path of the dependency, and strips off the root
directory of the write path. The remainder of the path is the relative path
of the dependency, which can then be used to find it in the installed package
files.
}
