% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_to_json.R
\name{chart_to_json}
\alias{chart_to_json}
\title{Convert chart object(s) to JSON}
\usage{
chart_to_json(x)
}
\arguments{
\item{x}{A single \code{highchartzero} object or a named list of \code{highchartzero}
objects.}
}
\value{
A \code{json} object from \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.
}
\description{
Takes a single \code{highchartzero} object or a named list of \code{highchartzero}
objects and returns a JSON representation of the object(s).
}
\examples{
data(example_charts)

chart_to_json(example_charts)

chart_1 = example_charts[[1]]
chart_to_json(chart_1)

}
