% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_webpage.R
\name{create_webpage}
\alias{create_webpage}
\title{Render webpage with chart}
\usage{
create_webpage(
  chart_objects,
  html_filename,
  root_dir = ".",
  page_title = "Chart",
  initial_chart = NULL,
  chart_picker_text = NULL,
  view = TRUE
)
}
\arguments{
\item{chart_objects}{Either a single \code{highcharterzero} object or a list of
\code{highcharterzero} objects.}

\item{html_filename}{Name of the saved html file.}

\item{root_dir}{Path to the webpage root directory}

\item{page_title}{The text for page's \code{title} element.}

\item{initial_chart}{The name of the list element with the initial chart to
show when the page is first loaded.}

\item{chart_picker_text}{A character vector of the names of the charts to
show in the dropdown selection element.}

\item{view}{View rendered page in browser. Ignored if not running
interactively.}
}
\description{
Render webpage with chart
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(example_charts)
create_webpage(
  example_charts,
  "example_charts.html",
  "Example Charts",
  "chart1",
  c("Chart 1", "Chart 2", "Chart 3")
)
\dontshow{\}) # examplesIf}
}
