% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_charts}
\alias{example_charts}
\title{Example Highcharts}
\format{
An object of class \code{list} of length 3.
}
\usage{
example_charts
}
\description{
A list of 3 example Highcharts. All 3 are \code{highchartzero} objects.
}
\keyword{datasets}
