% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chart_data.R
\name{export_chart_data}
\alias{export_chart_data}
\title{Write chart(s) JS to disk}
\usage{
export_chart_data(x, path)
}
\arguments{
\item{x}{A single \code{highchartzero} object or a named list of \code{highchartzero}
objects.}

\item{path}{The location of the JSON file.}
}
\description{
Converts a single \code{highchartzero} object or a named list of \code{highchartzero}
objects to JSON and writes the JSON to disk.
}
